add useContext for device state
