--- window-size-tracker-useEffect-useContext/src/index.js
@@ -1,13 +1,0 @@
-import React, { StrictMode } from "react";
-import ReactDOM from "react-dom";
-import { createRoot } from "react-dom/client";
-import App from "./App";
-
-const rootElement = document.getElementById("root");
-const root = createRoot(rootElement);
-
-root.render(
-  <StrictMode>
-    <App />
-  </StrictMode>
-);
