+++ window-size-tracker-useEffect/src/App.js
@@ -1,7 +1,8 @@
-import "./styles.css";
+// import "./styles.css";
 import React, { useState, useCallback } from "react";
 import WindowSizeComponent from "./WindowSizeComponent";
 import UpdateBgColor from "./UpdateBgColor";
+import DeviceContext from "./DeviceContext";
 
 function App() {
   const [device, setDevice] = useState("");
@@ -15,11 +16,10 @@ function App() {
   return (
     <div className="App">
       <h1>Window Size Tracker</h1>
-      <WindowSizeComponent
-        device={device}
-        onDeviceChange={handleDeviceChange}
-      />
-      <UpdateBgColor device={device} />
+      <DeviceContext.Provider value={device}>
+        <WindowSizeComponent onDeviceChange={handleDeviceChange} />
+        <UpdateBgColor />
+      </DeviceContext.Provider>
     </div>
   );
 }
