+++ window-size-tracker-useEffect/src/UpdateBgColor.jsx
@@ -1,7 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useContext } from 'react';
+import DeviceContext from './DeviceContext';
 
 // Function to update color depending on device
-function UpdateBgColor({ device }) {
+function UpdateBgColor() {
+  const device = useContext(DeviceContext);
   const [bgColor, setBgColor] = useState('');
   useEffect(() => {
     const color =
